Added AbstractPersistentCollectionFactory
